feat: move found instructions to action history
